pyproject.toml
[project]
name = "neo4j-tenant-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j>=5.28.2",
    "python-dotenv>=1.1.1",
    "wonderwords>=2.2.0",
]
